requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyMuPDF
pdfplumber
python-multipart
